+++ packages/expo-router/src/layouts/JSStack.tsx
@@ -25,6 +25,7 @@ export * from '../react-navigation/stack';
  * Creates the adapter props required to integrate the JavaScript stack with Expo Router.
  */
 export function unstable_createPropsForJSStack({
+  dispatch,
   dispatchSync,
   navigation,
   state,
@@ -33,6 +34,7 @@ export function unstable_createPropsForJSStack({
 >): StackNavigatorCreateProps {
   return {
     pop: makePopAction(dispatchSync, state.key),
+    removeRoutes: (routeNames) => dispatch({ type: 'REMOVE_ROUTES', payload: { routeNames } }),
     restoreRoute: makeRestoreRouteAction(dispatchSync, state),
     subscribePopToTopOnParentTabPress: () => subscribePopToTopOnParentTabPress(navigation, state),
   };